Settings now updates alarm times
