+++ app/src/main/java/alison/fivethingskotlin/Fragments/TimePreferenceFragment.kt
@@ -1,18 +1,16 @@
 package alison.fivethingskotlin.Fragments
 
+import alison.fivethingskotlin.util.NotificationScheduler
 import alison.fivethingskotlin.util.TimePreference
+import alison.fivethingskotlin.util.timeToString
 import android.content.Context
 import android.support.v7.preference.PreferenceDialogFragmentCompat
 import android.view.View
-import kotlinx.android.synthetic.main.pref_time_dialog.*
 import android.widget.TimePicker
-
-
-
 
 class TimePreferenceFragment: PreferenceDialogFragmentCompat() {
 
-    var timePicker: TimePicker? = null
+    private var timePicker: TimePicker? = null
 
     override fun onCreateDialogView(context: Context): View {
         timePicker = TimePicker(context)
@@ -34,8 +32,11 @@ class TimePreferenceFragment: PreferenceDialogFragmentCompat() {
             pref.hour = timePicker?.hour!!
             pref.minute = timePicker?.minute!!
 
-            val value = pref.timeToString(pref.hour, pref.minute)
-            if (pref.callChangeListener(value)) pref.persistStringValue(value)
+            val value = timeToString(pref.hour, pref.minute)
+            if (pref.callChangeListener(value)) {
+                pref.persistStringValue(value)
+                NotificationScheduler().setReminderNotification(context!!)
+            }
         }
     }
