+++ app/src/main/java/alison/fivethingskotlin/util/NotificationScheduler.kt
@@ -4,6 +4,7 @@ import android.app.AlarmManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
+import android.support.v7.preference.PreferenceManager
 import android.util.Log
 import java.util.*
 
@@ -17,7 +18,14 @@ class NotificationScheduler {
     private var alarmMgr: AlarmManager? = null
     private lateinit var alarmIntent: PendingIntent
 
-    fun setReminderNotification(context: Context, hour: Int, minute: Int) {
+    fun setReminderNotification(context: Context) {
+
+        val sharedPref = PreferenceManager.getDefaultSharedPreferences(context)
+        val prefTime = sharedPref.getString("pref_time", "default")
+        Log.d("blerg", "preftime: $prefTime")
+
+        val hour = parseHour(prefTime)
+        val minute = parseMinute(prefTime)
 
         val calendar: Calendar = Calendar.getInstance().apply {
             timeInMillis = System.currentTimeMillis()
